Feature/20241204 initial changes

+++ src/main/java/com/atradius/einvoice/ap/pdf/PdfCell.java
@@ -3,13 +3,15 @@ package com.atradius.einvoice.ap.pdf;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class PdfCell {
     public static final String LEFT = "Left";
     public static final String RIGHT = "Right";
-    private float width;
-    private float marginX;
-    private float marginY;
+    private BigDecimal width;
+    private BigDecimal marginX;
+    private BigDecimal marginY;
     private String alignment;
 }
